Fix tests

+++ vector.hpp
@@ -322,7 +322,7 @@ namespace ft
 	vector<T>::vector(vector::size_type n, const value_type &val) : m_size(n), m_capacity(n)
 	{
 		m_array = new value_type [n];
-		for (long i = 0; i < n; ++i)
+		for (long i = 0; i < static_cast<long>(n); ++i)
 			m_array[i] = val;
 	}
 
@@ -367,7 +367,7 @@ namespace ft
 			m_size = x.m_size;
 			m_capacity = x.m_capacity;
 
-			for (int i = 0; i < x.m_size; ++i)
+			for (long i = 0; i < static_cast<long>(x.m_size); ++i)
 				m_array[i] = x[i];
 		}
 		return *this;
@@ -409,16 +409,16 @@ namespace ft
 				if (n > 9)
 					m_capacity = n;
 				value_type *newArr = new value_type [m_capacity];
-				for (long i = 0; i < m_size; ++i)
+				for (long i = 0; i < static_cast<long>(m_size); ++i)
 					newArr[i] = m_array[i];
-				for (long i = m_size; i < n; ++i)
+				for (long i = m_size; i < static_cast<long>(n); ++i)
 					newArr[i] = val;
 				delete [] m_array;
 				m_array = newArr;
 			}
 			else
 			{
-				for (long i = m_size; i < n; ++i)
+				for (long i = m_size; i < static_cast<long>(n); ++i)
 					m_array[i] = val;
 			}
 			m_size = n;
@@ -510,7 +510,7 @@ namespace ft
 		{
 			m_capacity = getNearestPowerOfTwo(m_size + 1);
 			value_type *newArr = new value_type [m_capacity];
-			for (long i = 0; i < m_size; ++i)
+			for (long i = 0; i < static_cast<long>(m_size); ++i)
 				newArr[i] = m_array[i];
 			delete [] m_array;
 			m_array = newArr;
@@ -554,7 +554,7 @@ namespace ft
 			m_capacity = n;
 			m_array = new value_type [m_capacity];
 		}
-		for (long i = 0; i < n; ++i)
+		for (long i = 0; i < static_cast<long>(n); ++i)
 			m_array[i] = val;
 		m_size = n;
 	}
@@ -584,7 +584,7 @@ namespace ft
 		{
 			m_capacity = n;
 			value_type *newArr = new value_type [m_capacity];
-			for (long i = 0; i < m_size; ++i)
+			for (long i = 0; i < static_cast<long>(m_size); ++i)
 				newArr[i] = m_array[i];
 			delete [] m_array;
 			m_array = newArr;
@@ -670,7 +670,7 @@ namespace ft
 			m_capacity = getNearestPowerOfTwo(m_size + 1);
 			value_type *newArr = new value_type [m_capacity];
 			long i = m_size;
-			for (; i > (position - begin()); --i)
+			for (; i > static_cast<long>(position - begin()); --i)
 				newArr[i] = m_array[i - 1];
 			newArr[i] = val;
 			long newElementIndex = i;
@@ -699,9 +699,9 @@ namespace ft
 			m_capacity = getNearestPowerOfTwo(m_size + n + 1);
 			value_type *newArr = new value_type [m_capacity];
 			long i = m_size + n - 1;
-			for (; i >= (position - begin() + n); --i)
+			for (; i >= static_cast<long>(position - begin() + n); --i)
 				newArr[i] = m_array[i - n];
-			for (long j = 0; j < n; ++j)
+			for (long j = 0; j < static_cast<long>(n); ++j)
 			{
 				newArr[i] = val;
 				--i;
@@ -716,7 +716,7 @@ namespace ft
 			iterator it = end() + n - 1;
 			for (; it >= position + n; --it)
 				*it = *(it - n);
-			for (long i = 0; i < n; ++i)
+			for (long i = 0; i < static_cast<long>(n); ++i)
 			{
 				*position = val;
 				++position;
@@ -734,7 +734,7 @@ namespace ft
 			return;
 		size_type n = last - first;
 		value_type tmpArr[n];
-		for (long i = 0; i < n; ++i)
+		for (long i = 0; i < static_cast<long>(n); ++i)
 		{
 			tmpArr[i] = *first;
 			++first;
@@ -744,7 +744,7 @@ namespace ft
 			m_capacity = getNearestPowerOfTwo(m_size + n + 1);
 			value_type *newArr = new value_type [m_capacity];
 			long i = m_size + n - 1;
-			for (; i >= (position - begin() + n); --i)
+			for (; i >= static_cast<long>(position - begin() + n); --i)
 				newArr[i] = m_array[i - n];
 			for (long j = static_cast<long>(n) - 1; j >= 0; --j)
 			{
@@ -761,7 +761,7 @@ namespace ft
 			iterator it = end() + n - 1;
 			for (; it >= position + n; --it)
 				*it = *(it - n);
-			for (long i = 0; i < n; ++i)
+			for (long i = 0; i < static_cast<long>(n); ++i)
 			{
 				*position = tmpArr[i];
 				++position;
